# Conflicts:
#	api/app/Controllers/Init/Init.php
#	api/app/Controllers/Init/Languages.php
#	legacy/include/Notifications/NotificationManager.php
#	legacy/include/Notifications/NotificationPlugin.php
#	legacy/include/SugarEmailAddress/SugarEmailAddress.js
#	legacy/include/database/DBManager.php
#	legacy/include/language/en_us.lang.php
#	legacy/modules/Calendar/language/en_us.lang.php
#	legacy/modules/Employees/language/en_us.lang.php
#	legacy/modules/InboundEmail/language/en_us.lang.php
#	legacy/modules/KReports/vardefs.php
#	legacy/modules/Meetings/jsclass_scheduler.js
#	legacy/modules/MergeRecords/Step3.php
#	legacy/modules/Notes/vardefs.php
#	legacy/modules/OAuth2Tokens/language/en_us.lang.php
#	legacy/modules/ProspectLists/language/en_us.lang.php
#	legacy/themes/SuiteP/css/Mint/style.css
#	legacy/themes/SuiteP/css/Mint/style.css.map

+++ legacy/jssource/src_files/modules/Meetings/jsclass_scheduler.js
@@ -960,51 +960,53 @@ SugarWidgetScheduleRow.prototype.display = function () {
       } else {
          tr = this.thetable.insertRow( this.thetable.rows.length );
       }
-      tr.className = "schedulerAttendeeRow";
-      $( tr ).attr( 'data-id', this.focus_bean.fields.id );
-      // MintHCM #54195 #59793 Start
-      if ( this.focus_bean.module == 'Candidates' || this.focus_bean.module == 'Resources' ) {
-         $( tr ).attr( 'data-module', this.focus_bean.module );
-      } else {
-         $( tr ).attr( 'data-module', this.focus_bean.module + 's' );
+      if (tr) {
+        tr.className = "schedulerAttendeeRow";
+        $( tr ).attr( 'data-id', this.focus_bean.fields.id );
+        // MintHCM #54195 #59793 Start
+        if ( this.focus_bean.module == 'Candidates' || this.focus_bean.module == 'Resources' ) {
+          $( tr ).attr( 'data-module', this.focus_bean.module );
+        } else {
+          $( tr ).attr( 'data-module', this.focus_bean.module + 's' );
+        }
+        // MintHCM #54195 #59793 End
+        td = document.createElement( 'td' );
+        tr.appendChild( td );
+        //insertCell(tr.cells.length);
+
+        // icon + full name
+        td.scope = 'row';
+        var img = '<span class="suitepicon suitepicon-module-' + this.focus_bean.module.toLowerCase().replace( '_', '-' ) + '"></span>';
+
+        td.innerHTML = img;
+
+        td.innerHTML = td.innerHTML;
+
+        if ( this.focus_bean.fields.full_name )
+          td.innerHTML += ' ' + this.focus_bean.fields.full_name;
+        else
+          td.innerHTML += ' ' + this.focus_bean.fields.name;
+        // add freebusy tds here:
+        this.add_freebusy_nodes( tr );
+
+        // delete button
+        var td = document.createElement( 'td' );
+        tr.appendChild( td );
+        //var td = tr.insertCell(tr.cells.length);
+        td.className = 'schedulerAttendeeDeleteCell';
+        td.noWrap = true;
+        //CCL - Remove check to disallow removal of assigned user or current user
+        //if ( GLOBAL_REGISTRY.focus.fields.assigned_user_id != this.focus_bean.fields.id && GLOBAL_REGISTRY.current_user.fields.id != this.focus_bean.fields.id) {
+        td.innerHTML = '<a title="' + GLOBAL_REGISTRY['meeting_strings']['LBL_REMOVE']
+                + '" class="listViewTdToolsS1" style="text-decoration:none;" '
+                + 'href="javascript:SugarWidgetScheduleRow.deleteRow(\'' + this.focus_bean.fields.id + '\');">&nbsp;'
+                + '<img src="index.php?entryPoint=getImage&themeName=' + SUGAR.themes.theme_name + '&imageName=delete_inline.gif" '
+                + 'align="absmiddle" alt="' + GLOBAL_REGISTRY['meeting_strings']['LBL_REMOVE'] + '" border="0"> '
+                + GLOBAL_REGISTRY['meeting_strings']['LBL_REMOVE'] + '</a>';
+        //}
+        this.element = tr;
+        this.element_index = this.thetable.rows.length - 1;
       }
-      // MintHCM #54195 #59793 End
-      td = document.createElement( 'td' );
-      tr.appendChild( td );
-      //insertCell(tr.cells.length);
-
-      // icon + full name
-      td.scope = 'row';
-      var img = '<span class="suitepicon suitepicon-module-' + this.focus_bean.module.toLowerCase().replace( '_', '-' ) + '"></span>';
-
-      td.innerHTML = img;
-
-      td.innerHTML = td.innerHTML;
-
-      if ( this.focus_bean.fields.full_name )
-         td.innerHTML += ' ' + this.focus_bean.fields.full_name;
-      else
-         td.innerHTML += ' ' + this.focus_bean.fields.name;
-      // add freebusy tds here:
-      this.add_freebusy_nodes( tr );
-
-      // delete button
-      var td = document.createElement( 'td' );
-      tr.appendChild( td );
-      //var td = tr.insertCell(tr.cells.length);
-      td.className = 'schedulerAttendeeDeleteCell';
-      td.noWrap = true;
-      //CCL - Remove check to disallow removal of assigned user or current user
-      //if ( GLOBAL_REGISTRY.focus.fields.assigned_user_id != this.focus_bean.fields.id && GLOBAL_REGISTRY.current_user.fields.id != this.focus_bean.fields.id) {
-      td.innerHTML = '<a title="' + GLOBAL_REGISTRY['meeting_strings']['LBL_REMOVE']
-              + '" class="listViewTdToolsS1" style="text-decoration:none;" '
-              + 'href="javascript:SugarWidgetScheduleRow.deleteRow(\'' + this.focus_bean.fields.id + '\');">&nbsp;'
-              + '<img src="index.php?entryPoint=getImage&themeName=' + SUGAR.themes.theme_name + '&imageName=delete_inline.gif" '
-              + 'align="absmiddle" alt="' + GLOBAL_REGISTRY['meeting_strings']['LBL_REMOVE'] + '" border="0"> '
-              + GLOBAL_REGISTRY['meeting_strings']['LBL_REMOVE'] + '</a>';
-      //}
-      this.element = tr;
-      this.element_index = this.thetable.rows.length - 1;
    }, null, this );
 }
